Implement support for Glint language server v2
Zed just got support for forwarding tsserver messages in the latest
preview. The caveat it only works if the language server is named
`vue-language-server`. I've asked if they are willing to support other
language-servers in the same fasion but I haven't heared back.

The implementation is just Vue's language server, but replaced vue's
packages with Glint's packages.

Requires these settings to be added to the languages in your Zed
settings

```
    "Glimmer (TypeScript)": {
      "language_servers": ["vtsls", "..."],
      "prettier": {
        "allowed": true,
        "parser": "ember-template-tag",
        "plugins": ["prettier-plugin-ember-template-tag"]
      }
    },
    "Glimmer (JavaScript)": {
      "language_servers": ["vtsls", "..."],
      "prettier": {
        "allowed": true,
        "parser": "ember-template-tag",
        "plugins": ["prettier-plugin-ember-template-tag"]
      }
    },
```